arithmetics example: moved util method to dedicated module to avoid importing 'fs' via 'main-browser', added explicit package exports
